+++ libs/community-contract/contract/src/lib.rs
@@ -1,9 +1,9 @@
-pub mod account_info;
+// pub mod account_info;
 pub mod admin;
 pub mod nft;
 pub mod errors;
 
-use crate::account_info::AccountInfos;
+// use crate::account_info::AccountInfos;
 use crate::admin::Admin;
 use crate::errors::throw_error;
 use crate::nft::NFTFunc;
@@ -93,51 +93,51 @@ impl Admin for Community {
     }
 }
 
-#[near_bindgen]
-impl AccountInfos for Community {
-    fn get_account_nonce(&self, secp_pk: SecpPK) -> i32 {
-        self.get_account_info(secp_pk)
-            .map(|account_info| account_info.nonce)
-            .unwrap_or(0)
-    }
+// #[near_bindgen]
+// impl AccountInfos for Community {
+//     fn get_account_nonce(&self, secp_pk: SecpPK) -> i32 {
+//         self.get_account_info(secp_pk)
+//             .map(|account_info| account_info.nonce)
+//             .unwrap_or(0)
+//     }
 
-    fn get_account_id(&self, secp_pk: SecpPK) -> Option<AccountId> {
-        self.get_account_info(secp_pk)
-            .map(|account_info| account_info.account_id)
-    }
+//     fn get_account_id(&self, secp_pk: SecpPK) -> Option<AccountId> {
+//         self.get_account_info(secp_pk)
+//             .map(|account_info| account_info.account_id)
+//     }
 
-    fn set_account_info(
-        &mut self,
-        user_id: String,
-        secp_pk: SecpPK,
-        secp_sig_s: Vec<u8>,
-        new_account_id: AccountId,
-    ) {
-        if !is_valid_account_id(new_account_id.as_bytes()) {
-            throw_error(crate::errors::INVALID_ACCOUNT_ID);
-        }
+//     fn set_account_info(
+//         &mut self,
+//         user_id: String,
+//         secp_pk: SecpPK,
+//         secp_sig_s: Vec<u8>,
+//         new_account_id: AccountId,
+//     ) {
+//         if !is_valid_account_id(new_account_id.as_bytes()) {
+//             throw_error(crate::errors::INVALID_ACCOUNT_ID);
+//         }
 
-        if !(predecessor_account_id() == new_account_id
-            || self.admins.contains(&predecessor_account_id()))
-        {
-            throw_error(crate::errors::UNAUTHORIZED);
-        }
+//         if !(predecessor_account_id() == new_account_id
+//             || self.admins.contains(&predecessor_account_id()))
+//         {
+//             throw_error(crate::errors::UNAUTHORIZED);
+//         }
 
-        let (secp_pk_internal, nonce) = self.verify_sig(user_id, secp_pk, secp_sig_s);
-        self.account_infos.insert(
-            &secp_pk_internal,
-            &AccountInfo {
-                account_id: new_account_id,
-                nonce: nonce + 1,
-            },
-        );
-    }
+//         let (secp_pk_internal, nonce) = self.verify_sig(user_id, secp_pk, secp_sig_s);
+//         self.account_infos.insert(
+//             &secp_pk_internal,
+//             &AccountInfo {
+//                 account_id: new_account_id,
+//                 nonce: nonce + 1,
+//             },
+//         );
+//     }
 
-    fn delete_account_info(&mut self, user_id: String, secp_pk: SecpPK, secp_sig_s: Vec<u8>) {
-        let (secp_pk_internal, _) = self.verify_sig(user_id, secp_pk, secp_sig_s);
-        // TODO: this leaves vulnrebaility to replay attacks. If an account is made, deleted, and made again,
-        // The nonce resets to 0. Please see https://github.com/bafnetwork/baf-wallet-v2/issues/32
-        self.account_infos.remove(&secp_pk_internal);
-    }
-}
+//     fn delete_account_info(&mut self, user_id: String, secp_pk: SecpPK, secp_sig_s: Vec<u8>) {
+//         let (secp_pk_internal, _) = self.verify_sig(user_id, secp_pk, secp_sig_s);
+//         // TODO: this leaves vulnrebaility to replay attacks. If an account is made, deleted, and made again,
+//         // The nonce resets to 0. Please see https://github.com/bafnetwork/baf-wallet-v2/issues/32
+//         self.account_infos.remove(&secp_pk_internal);
+//     }
+// }
