+++ libs/community-contract/src/lib/community-contract.ts
@@ -10,20 +10,21 @@ import {
 import { NearAccountID } from '@baf-wallet/near';
 import { pkToArray } from '@baf-wallet/crypto';
 
+// TODO: the rest of the interface
 interface CommunityContract {
-  getAccountId: (pk: PublicKey<secp256k1>) => Promise<NearAccountID | null>;
-  getAccountNonce: (secp_pk: PublicKey<secp256k1>) => Promise<string>;
-  setAccountInfo: (
-    secp_pk: PublicKey<secp256k1>,
-    user_id: string,
-    secp_sig_s: RustEncodedSecpSig,
-    new_account_id: NearAccountID
-  ) => Promise<void>;
-  deleteAccountInfo: (
-    secp_pk: PublicKey<secp256k1>,
-    user_id: string,
-    secp_sig_s: RustEncodedSecpSig
-  ) => Promise<void>;
+  // getAccountId: (pk: PublicKey<secp256k1>) => Promise<NearAccountID | null>;
+  // getAccountNonce: (secp_pk: PublicKey<secp256k1>) => Promise<string>;
+  // setAccountInfo: (
+  //   secp_pk: PublicKey<secp256k1>,
+  //   user_id: string,
+  //   secp_sig_s: RustEncodedSecpSig,
+  //   new_account_id: NearAccountID
+  // ) => Promise<void>;
+  // deleteAccountInfo: (
+  //   secp_pk: PublicKey<secp256k1>,
+  //   user_id: string,
+  //   secp_sig_s: RustEncodedSecpSig
+  // ) => Promise<void>;
   get_default_nft_contract: () => Promise<string>;
 }
 
@@ -38,7 +39,7 @@ export async function setCommunityContract(
 
 export function getCommunityContract(): CommunityContract {
   if (communityContract) return communityContract;
-  throw BafError.UnintCommunityContract();
+  throw BafError.UninitGlobalContract();
 }
 
 async function buildCommunityContract(
@@ -46,14 +47,14 @@ async function buildCommunityContract(
 ): Promise<CommunityContract> {
   const contract = new Contract(account, ContractConfig.contractName, {
     viewMethods: [
-      'get_account_id',
-      'get_account_nonce',
+      // 'get_account_id',
+      // 'get_account_nonce',
       'get_admins',
       'get_default_nft_contract',
     ],
     changeMethods: [
-      'set_account_info',
-      'delete_account_info',
+      // 'set_account_info',
+      // 'delete_account_info',
       'add_admins',
       'remove_admins',
       'set_default_nft_contract',
@@ -66,29 +67,29 @@ async function buildCommunityContract(
      * Below are override functions for the calls
      * Find the contract code in libs/community-contract/contract
      */
-    getAccountId: async (pk) => {
-      const ret = await (contract as any).get_account_id({
-        secp_pk: pkToArray(pk),
-      });
-      if (!ret || ret === '') return null;
-      else return ret as NearAccountID;
-    },
-    getAccountNonce: (pk) =>
-      (contract as any).get_account_nonce({
-        secp_pk: pkToArray(pk),
-      }) as Promise<string>,
-    setAccountInfo: (pk, user_id, secp_sig_s, new_account_id) =>
-      (contract as any).set_account_info({
-        user_id,
-        secp_pk: pkToArray(pk),
-        secp_sig_s: [...secp_sig_s],
-        new_account_id,
-      }),
-    deleteAccountInfo: (pk, user_id, secp_sig_s) =>
-      (contract as any).delete_account_info({
-        user_id,
-        secp_pk: pkToArray(pk),
-        secp_sig_s: [...secp_sig_s],
-      }),
+    // getAccountId: async (pk) => {
+    //   const ret = await (contract as any).get_account_id({
+    //     secp_pk: pkToArray(pk),
+    //   });
+    //   if (!ret || ret === '') return null;
+    //   else return ret as NearAccountID;
+    // },
+    // getAccountNonce: (pk) =>
+    //   (contract as any).get_account_nonce({
+    //     secp_pk: pkToArray(pk),
+    //   }) as Promise<string>,
+    // setAccountInfo: (pk, user_id, secp_sig_s, new_account_id) =>
+    //   (contract as any).set_account_info({
+    //     user_id,
+    //     secp_pk: pkToArray(pk),
+    //     secp_sig_s: [...secp_sig_s],
+    //     new_account_id,
+    //   }),
+    // deleteAccountInfo: (pk, user_id, secp_sig_s) =>
+    //   (contract as any).delete_account_info({
+    //     user_id,
+    //     secp_pk: pkToArray(pk),
+    //     secp_sig_s: [...secp_sig_s],
+    //   }),
   };
 }
