--- libs/community-contract/src/lib/constants.ts
@@ -1,4 +1,0 @@
-export const CommunityContractConstants = {
-  // :)
-  beneficiaryId: 'lev.near',
-};
