+++ libs/errors/src/lib/errors.ts
@@ -23,9 +23,13 @@ export namespace BafError {
   export const UnsupportedKeyCurve = (supportedCurve: string) =>
     emitError(`Only ${supportedCurve} keys are supported`);
   export const BlockedByCors = () => emitError('Not allowed by CORS policy');
+  export const UninitGlobalContract = () =>
+    emitError(
+      'Global Contract is not initialized yet, please call setGlobalContract'
+    );
   export const UnintCommunityContract = () =>
     emitError(
-      'BAF Contract is not initialized yet, please call setCommunityContract'
+      'Community Contract is not initialized yet, please call setCommunityContract'
     );
   export const UnsupportedToken = (tokenName: string) =>
     emitError(`Unsupported token ${tokenName}`);
