+++ libs/global-contract/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["@nrwl/web/babel"]
+}
