+++ libs/global-contract/README.md
@@ -1,0 +1,7 @@
+# global-contract
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test global-contract` to execute the unit tests via [Jest](https://jestjs.io).
