+++ libs/global-contract/config.json
@@ -1,0 +1,1 @@
+{"contractName": "dev-1623353929808-71316364508823"}
