+++ libs/global-contract/contract/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "global-contract"
+version = "0.1.0"
+authors = ["Lev Stambler <levstamb@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+libsecp256k1 = "0.3.5"
+near-sdk = "3.1.0"
