+++ libs/global-contract/contract/build.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+RUSTFLAGS='-C link-arg=-s' cargo build --target wasm32-unknown-unknown --release
