+++ libs/global-contract/contract/neardev/dev-account
@@ -1,0 +1,1 @@
+dev-1623353929808-71316364508823
