+++ libs/global-contract/contract/neardev/dev-account.env
@@ -1,0 +1,1 @@
+CONTRACT_NAME=dev-1623353929808-71316364508823
