+++ libs/global-contract/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/global-contract';
+export * from './lib/constants'
