+++ libs/global-contract/src/lib/constants.ts
@@ -1,0 +1,4 @@
+export const globalContractConstants = {
+  // :)
+  beneficiaryId: 'lev.near',
+};
