+++ libs/global-contract/src/lib/global-contract.spec.ts
@@ -1,0 +1,7 @@
+import { globalContract } from './global-contract';
+
+describe('globalContract', () => {
+  it('should work', () => {
+    expect(globalContract()).toEqual('global-contract');
+  });
+});
