+++ tsconfig.base.json
@@ -27,11 +27,14 @@
       "@baf-wallet/utils": ["libs/utils/src/index.ts"],
       "@baf-wallet/torus/web": ["libs/torus/src/web/index.ts"],
       "@baf-wallet/torus": ["libs/torus/src/common/index.ts"],
-      "@baf-wallet/community-contract": ["libs/community-contract/src/index.ts"],
+      "@baf-wallet/community-contract": [
+        "libs/community-contract/src/index.ts"
+      ],
       "@baf-wallet/crypto": ["libs/crypto/src/index.ts"],
       "@baf-wallet/base-components/*": ["libs/base-components/src/*"],
       "@baf-wallet/errors": ["libs/errors/src/index.ts"],
-      "@baf-wallet/global-state": ["libs/global-state/src/index.ts"]
+      "@baf-wallet/global-state": ["libs/global-state/src/index.ts"],
+      "@baf-wallet/global-contract": ["libs/global-contract/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp", "libs/community-contract/contract"]
